+++ contracts/mocks/DinngoProxyMock.sol
@@ -1,0 +1,44 @@
+pragma solidity ^0.4.24;
+
+import "../DinngoProxy.sol";
+
+contract DinngoProxyMock is DinngoProxy {
+    event AddUser(uint256 userID, address indexed user);
+    event AddToken(uint256 tokenID, address indexed token);
+    event Deposit(address token, address indexed user, uint256 amount, uint256 balance);
+    event Withdraw(address token, address indexed user, uint256 amount, uint256 balance);
+    event Trade(
+        address indexed user,
+        bool isBuy,
+        address indexed tokenTarget,
+        uint256 amountTarget,
+        address indexed tokenTrade,
+        uint256 amountTrade
+    );
+    event Lock(address indexed user, uint256 lockTime);
+    event Unlock(address indexed user);
+
+    constructor(address dinngoWallet, address dinngoToken, address impl)
+        DinngoProxy(dinngoWallet, dinngoToken, impl)
+        public
+    {
+    }
+
+    function setUserBalance(address user, address token, uint256 amount) external {
+        balances[token][user] = amount;
+    }
+
+    function setUser(uint256 userID, address user, uint8 rank) external {
+        userID_Address[userID] = user;
+        userRanks[user] = rank;
+    }
+
+    function setToken(uint256 tokenID, address token, uint8 rank) external {
+        tokenID_Address[tokenID] = token;
+        tokenRanks[token] = rank;
+    }
+
+    function fillOrder(bytes32 hash, uint256 amount) external {
+        orderFills[hash] = orderFills[hash].add(amount);
+    }
+}