+++ test/DepositProxy.test.js
@@ -5,7 +5,8 @@ import { inLogs } from 'openzeppelin-solidity/test/helpers/expectEvent';
 const BigNumber = web3.BigNumber;
 const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000';
 
-const DinngoMock = artifacts.require('DinngoMock');
+const Dinngo = artifacts.require('Dinngo');
+const DinngoProxyMock = artifacts.require('DinngoProxyMock');
 const SimpleToken = artifacts.require('SimpleToken');
 
 require('chai')
@@ -13,10 +14,10 @@ require('chai')
     .use(require('chai-bignumber')(BigNumber))
     .should();
 
-/*
 contract('Deposit', function ([_, user, owner, tokenWallet, tokenContract]) {
     beforeEach(async function () {
-        this.Dinngo = await DinngoMock.new(tokenWallet, tokenContract, { from: owner });
+        this.DinngoImpl = await Dinngo.new(tokenWallet, tokenContract, { from: owner });
+        this.Dinngo = await DinngoProxyMock.new(tokenWallet, tokenContract, this.DinngoImpl.address, { from: owner });
     });
 
     describe('ether', function () {
@@ -83,4 +84,3 @@ contract('Deposit', function ([_, user, owner, tokenWallet, tokenContract]) {
         });
     });
 });
-*/