+++ test/SettleProxy.test.js
@@ -1,11 +1,11 @@
 import { ether } from 'openzeppelin-solidity/test/helpers/ether';
-import expectThrow from 'openzeppelin-solidity/test/helpers/expectThrow';
+import { reverting } from 'openzeppelin-solidity/test/helpers/shouldFail';
 
 const BigNumber = web3.BigNumber;
 const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000';
 
-const DinngoMock = artifacts.require('DinngoMock');
-
+const Dinngo = artifacts.require('Dinngo');
+const DinngoProxyMock = artifacts.require('DinngoProxyMock');
 const should = require('chai')
     .use(require('chai-bignumber')(BigNumber))
     .use(require('chai-as-promised'))
@@ -14,7 +14,8 @@ const should = require('chai')
 contract('Settle', function([_, user1, user2, user3, user4, user5, owner, dinngoWallet, DGO, token]) {
     const BALANCE = ether(1000);
     beforeEach(async function() {
-        this.Dinngo = await DinngoMock.new(dinngoWallet, DGO, { from: owner });
+        this.DinngoImpl = await Dinngo.new(dinngoWallet, DGO, { from: owner });
+        this.Dinngo = await DinngoProxyMock.new(dinngoWallet, DGO, this.DinngoImpl.address, { from: owner });
         await this.Dinngo.setUser(11, user1, 1);
         await this.Dinngo.setUser(12, user2, 1);
         await this.Dinngo.setUser(13, user3, 1);
@@ -39,51 +40,6 @@ contract('Settle', function([_, user1, user2, user3, user4, user5, owner, dinngo
         await this.Dinngo.setUserBalance(dinngoWallet, token, BALANCE);
         await this.Dinngo.setUserBalance(dinngoWallet, ZERO_ADDRESS, BALANCE);
         await this.Dinngo.setUserBalance(dinngoWallet, DGO, BALANCE);
-    });
-
-    describe('process maker order', function() {
-        // user ID: 11
-        // target token: 0
-        // target amount: 1 ether
-        // trade token: 11
-        // trade amount: 100 ether
-        // config: 1+2+16 (isBuy = true; isFeeMain = true; Gas price = 4 Gwei)
-        // fee: 1
-        // nonce: 1
-        const order1 = "0x24f3a54a1b754191b87333031b45fabfd20139056c391386597a7b3598856dd1e210212a1d3790c95bb3e56bfac578ab59479aef335ae48885ec48295fd510f80000000000000000000000000000000000000000000000000000038d7ea4c680000000000000000000000000000000000000000000000000000de0b6b3a764000000000001030000000000000000000000000000000000000000000000056bc75e2d63100000000b0000000000000000000000000000000000000000000000000de0b6b3a764000000000000000b";
-        const tokenTarget1 = ZERO_ADDRESS;
-        const tokenTrade1 = token;
-        const amountTarget1 = ether(1);
-        const amountTrade1 = ether(100);
-        // user ID: 12
-        // target token: 11
-        // target amount: 10 ether
-        // trade token: 0
-        // trade amount: 0.1 ether
-        // config: 2+16 (isBuy = false; isFeeMain = true; Gas price = 4 Gwei)
-        // fee: 0.1
-        // nonce: 2
-        const order2 = "0x5cf7ebade4232d752b01f797193d7d9ea60de04083ba43d9ae36fc0be5709f2b3620137ce74643e47f0c0225a43b433d48fc0e44dbeaf53f92bea446a1c003210100000000000000000000000000000000000000000000000000038d7ea4c68000000000000000000000000000000000000000000000000000016345785d8a00000000000202000000000000000000000000000000000000000000000000016345785d8a000000000000000000000000000000000000000000000000000000008ac7230489e80000000b0000000c";
-        const tokenTarget2 = token;
-        const tokenTrade2 = ZERO_ADDRESS;
-        const amountTarget2 = ether(10);
-        const amountTrade2 = ether(0.1);
-        it('Taker target greater than buying maker trade', async function() {
-            const restAmountTarget = ether(130);
-            const { logs } = await this.Dinngo.processMakerMock(order1, restAmountTarget);
-            const event = logs.find(e => e.event === "TestMaker");
-            should.exist(event);
-            event.args.fillAmountTrade.should.be.bignumber.eq(amountTarget1);
-            event.args.restAmountTarget.should.be.bignumber.eq(restAmountTarget.minus(amountTrade1));
-        });
-        it('Taker target lesser than buying maker trade', async function() {
-            const restAmountTarget = ether(80);
-            const { logs } = await this.Dinngo.processMakerMock(order1, restAmountTarget);
-            const event = logs.find(e => e.event === "TestMaker");
-            should.exist(event);
-            event.args.fillAmountTrade.should.be.bignumber.eq(amountTarget1.mul(restAmountTarget).div(amountTrade1).toFixed(0));
-            event.args.restAmountTarget.should.be.bignumber.eq(ether(0));
-        });
     });
 
     describe('settle', function() {
@@ -207,12 +163,10 @@ contract('Settle', function([_, user1, user2, user3, user4, user5, owner, dinngo
             amount1.should.be.bignumber.eq(amountTrade2);
             amount2.should.be.bignumber.eq(amountTarget2);
         });
-
         it('Normal count gas 1-1', async function() {
             const receipt = await this.Dinngo.settle(orders1_2, { from: owner });
             console.log(receipt.receipt.gasUsed);
         });
-
         it('Normal count gas 1-2', async function() {
             const receipt = await this.Dinngo.settle(orders1_2_3, { from: owner });
             console.log(receipt.receipt.gasUsed);
@@ -368,26 +322,36 @@ contract('Settle', function([_, user1, user2, user3, user4, user5, owner, dinngo
 
         it('taker invalid', async function() {
             await this.Dinngo.removeUser(user2, { from: owner });
-            await expectThrow(this.Dinngo.settle(orders1_2, { from: owner }));
+            await reverting(this.Dinngo.settle(orders1_2, { from: owner }));
         });
 
         it('maker invalid', async function() {
             await this.Dinngo.removeUser(user1, { from: owner });
-            await expectThrow(this.Dinngo.settle(orders1_2, { from: owner }));
+            await reverting(this.Dinngo.settle(orders1_2, { from: owner }));
         });
 
         it('taker invalid', async function() {
             await this.Dinngo.removeUser(user2, { from: owner });
-            await expectThrow(this.Dinngo.settle(orders1_2, { from: owner }));
+            await reverting(this.Dinngo.settle(orders1_2, { from: owner }));
         });
 
         it('maker invalid', async function() {
             await this.Dinngo.removeUser(user1, { from: owner });
-            await expectThrow(this.Dinngo.settle(orders1_2, { from: owner }));
+            await reverting(this.Dinngo.settle(orders1_2, { from: owner }));
+        });
+
+        it('taker order filled', async function() {
+            await this.Dinngo.fillOrder(hash1, amountTarget1);
+            await reverting(this.Dinngo.settle(orders1_2, { from: owner }));
+        });
+
+        it('maker order filled', async function() {
+            await this.Dinngo.fillOrder(hash2, amountTarget2);
+            await reverting(this.Dinngo.settle(orders1_2, { from: owner }));
         });
 
         it('Not owner', async function() {
-            await expectThrow(this.Dinngo.settle(orders1_2));
+            await reverting(this.Dinngo.settle(orders1_2));
         });
     });
 });