+++ truffle.js
@@ -16,6 +16,22 @@ require('babel-register')({
 });
 require('babel-polyfill');
 module.exports = {
-  // See <http://truffleframework.com/docs/advanced/configuration>
-  // to customize your Truffle configuration!
+    networks: {
+        development: {
+            host: "localhost",
+            port: 8242,
+            network_id: "*", // Match any network id
+            gasPrice: 0x01      // <-- Use this low gas price
+        },
+        integration: {
+            host: "integration.my-dapp.com",
+            port: 80,
+            network_id: "*"
+        }
+    },
+    compilers: {
+        solc: {
+            version: "0.4.25"
+        }
+    }
 };