+++ CHANGELOG.md
@@ -3,6 +3,23 @@ All notable changes to this project will be documented in this file.
 
 ## [unreleased]
 
+## [0.2.1] - 2018-12-11
+### Added
+- Added the solc version in truffle.js
+- Upgrade openzeppelin-solidity to 2.0.0
+- Continous Integration
+- Make locking process time modifiable
+
+### Fixed
+- Fix test to support openzeppelin-solidity 2.0.0
+- Adding removed user/token under same id should succeed
+- Orderfills is not checked after the update
+- Revert when settle does not behave normally
+
+## [0.2.0] - 2018-11-28
+### Added
+- Implement proxy structure
+
 ## [0.1.1] - 2018-11-29
 ### Fixed
 - Fix the bug caused by importing openzeppelin test helper